feat: use trainer as seperate py module
