+++ tests/integration/ranker_training/offline/yaml/train.yml
@@ -1,10 +1,11 @@
-!RankerTrainer
+!RandomWeightedRankerTrainer
 with:
   params:
     metric: ndcg
     objective: lambdarank
 metas:
   workspace: $JINA_TEST_TRAIN_OFFLINE
+  py_modules: trainer.py
 requests:
   on:
     TrainRequest:
